Check comments in FlyDecider.cs
im going to have to talk to Oldenburg about the right way to do this. Comments on FlyDecider should explain pretty well what I'm going for. Two errors, for the same damn line, say it's expecting a ";" and a "}". So im going to call it here until I can get some advice from the professor.