Fixed UI, added a new title scene +more
the fly will now instantiate with a rotation so it faces forwards.

the fly will now randomly pick a direction if the list is empty (some issues with this, but it prevents bugs/ softlocks)

changed the fly spawn height back to 0 since it was hard to understand the game state due to this.  I raised its hitbox by 1 to compensate.

removed the shadow from the fly, since it was also making it harder to interpret the board while the fly was raised